git-svn-id: svn://svn.code.sf.net/p/mcapi/code@221 3a57ee49-f89d-46b9-a3c0-71fdb2aa5dce

+++ src/TestCapi/Test_CapiCommon.cpp
@@ -55,9 +55,9 @@ void TestCase_GetCacheAlignedAddr(void)
 void TestCase_AtomicCAS(void)
 {
     int ret;
-    int val;
-    int oldval;
-    int newval;
+    LONG val;
+    LONG oldval;
+    LONG newval;
 
     val = 100;
     oldval = val;
@@ -66,5 +66,5 @@ void TestCase_AtomicCAS(void)
 
     ret = AtomicCAS(&val, newval, oldval);
 
-    printf("AtomicCAS(), oldval = %d, val = %d, ret = %d\n", oldval, val, ret);
+    printf("AtomicCAS(), oldval = %ld, val = %ld, ret = %d\n", oldval, val, ret);
 }
